Change `w` suffix to `-gui`
This way it can't possibly be a module name, so we don't need to inspect
whether we're running with pythonw before we strip it off.

Therefore we can unconditionally strip it off, and then running
`<module>-gui` on Linux won't mysteriously crash, which is the next best
thing to getting rid of these redundant gui scripts on unix altogether.