Add .cpp extension to hook into window creation
This allows us to set the AppUserModelID on all new windows without the
cooperation of the GUI toolkit, and working around the limitation that
SetCurrentProcessExplicitAppUserModelID doesn't work to associate the
taskbar items in Python from the Microsoft Store.

Since this requires compiling an extension, need to work out how to
build wheels that are platform-aware. Right now it's just an if
statement in setup.py.

To build from source you'll need a Microsoft visual C++ compiler.
See here: https://wiki.python.org/moin/WindowsCompilers

This addresses issue #1