Fix failure to install start menu items in Microsoft Store Python
This was caused by the filepaths for the shortcuts being redirected to a
private copy of %APPDATA% specific to the Microsoft Store Python app,
which windows does not look in to populate the start menu.

Although it's a filthy hack, we now detect when this has happened, and
use subprocesses running standard windows command line tools to copy the
file to the actual appdata directory where the start menu may pick it
up. The subprocesses are not subject to the redirect, so this works.

Fixes #3