Don't set argv[0] on Windows
It breaks the app relaunching itself via sys.argv, and isn't needed for
appid stuff. Don't set it on Linux unless it exists, for the same
reason. If it doesn't exist then the app is not installed properly, so
breaking appid stuff is fine - better to keep sys.argv as something that
can actually run the program still.