Fix #13 by appending `.exe` to launcher script path
It seems that shortcuts must include `.exe` in their `TargetPath`, but
that Windows often appends this extension automatically. However, when
the `TargetPath` already contains a `.`, then Windows does not
automatically append `.exe` (presumably thinking the target already has
an extension). However in our case the dot may be from a dotted
submodule name, so we still need `.exe` to be appended.