Fix an issue with recent versions of setuptools/pip
in which editable installs a) aren't in sys.path and therefore can't be
found using PathFinder, and b) don't have .egg-link files associated
with them.

Instead, just use `importlib.util.find_spec` to find the import path of
the module. This is the canonical way to do this, and the only reason I
was avoiding it was in order to not hit the problem of accidentally
using the current working directory as an import search path. However,
the current working directory is only prepended to `sys.path` when
running `python -m module` or `python` by itself, so this shouldn't be
a problem for the desktop-app launcher. That was probably overcautious
and resulted in fragile code.

Also, instead of looking for `.egg-link` files, we can instead look for
entries in `sys.modules` that contain information about editable
installs. These are stub modules that would live in e.g.
`site-packages`, so by looking up their `__file__`, we can tell what
import location they correspond to and thus find the corresponding
scripts directory correctly.